Koordy float, uogólnione update'y, śmierć Wszystkie współrzędne związane z położeniem pikselowym mają  typ float Update'y stanu gracza i wrogów zostały uogólnione Kontakt z bombą jest teraz liczony wg tyknięcia zegara gry, a nie sekund Gracz może teraz umrzeć od bomby,

+++ bomberman/headers/actor.h
@@ -15,9 +15,9 @@ vx[0]----0---> vx[1]
 
 typedef struct S_Actor
 {
-    double x, y;
+    float x, y;
     bool vx[2], vy[2];
-    double speed;
+    float speed;
     bool alive;
     enum Direction dir;
     ALLEGRO_BITMAP *bmp;
@@ -27,7 +27,7 @@ typedef struct S_Actor
 
 // ======================== actor_essentials.c =========================
 
-Actor *createActor( double x, double y, double speed, enum Direction dir, ALLEGRO_BITMAP *bmp );
+Actor *createActor( float x, float y, float speed, enum Direction dir, ALLEGRO_BITMAP *bmp );
 void destroyActor( Actor **actor );
 void destroyActors( Actor ** enemies, int enemy_num );
