+++ bomberman/src/ai/ai_essentials.c
@@ -27,7 +27,7 @@ void destroyAIModules( AIModule ** modules, int module_num )
         destroyAIModule( &modules[i] );
 }
 
-Actor *createAIActor( Path path, double speed, enum Direction dir, ALLEGRO_BITMAP *bmp )
+Actor *createAIActor( Path path, float speed, enum Direction dir, ALLEGRO_BITMAP *bmp )
 {
     return createActor( path.steps[0].dest_x, path.steps[0].dest_y, speed, dir, bmp );
 }
